Set custom domain for documentation site
The 'documentation-deploy' flow overwrites everything in the 'gh-pages' branch
every time there is a push to the master branch - that is why files like CNAME
go missing. There is a setting on the action that generates the site that should
set a custom domain name, which has been set appropriately.

Upgraded the checkout action. Removed commented code and todo.